refactor files structure of impala algorithm; seperate numpy utils from utils
