+++ parl/utils/np_utils.py
@@ -12,8 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from parl.algorithms.ddpg import *
-from parl.algorithms.dqn import *
-from parl.algorithms.policy_gradient import *
-from parl.algorithms.ppo import *
-from parl.algorithms.impala.impala import *
+import numpy as np
+
+__all__ = ['np_softmax', 'np_cross_entropy']
+
+def np_softmax(logits):
+    return np.exp(logits) / np.sum(np.exp(logits), axis=-1, keepdims=True)
+
+
+def np_cross_entropy(probs, labels):
+    if labels.shape[-1] == 1:
+        # sparse label
+        n_classes = probs.shape[-1]
+        result_shape = list(labels.shape[:-1]) + [n_classes]
+        labels = np.eye(n_classes)[labels.reshape(-1)]
+        labels = labels.reshape(result_shape)
+
+    return -np.sum(labels * np.log(probs), axis=-1, keepdims=True)
